CHORE #264 : config 패키지 생성 후 디렉토리 구조 이동
