+++ src/main/java/com/lokoko/global/auth/provider/google/config/GoogleProperties.java
@@ -1,10 +1,10 @@
-package com.lokoko.global.auth.provider.tiktok;
+package com.lokoko.global.auth.provider.google.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-@ConfigurationProperties(prefix = "tiktok")
-public record TikTokProperties(
-        String clientKey,
+@ConfigurationProperties(prefix = "google")
+public record GoogleProperties(
+        String clientId,
         String clientSecret,
         String redirectUri,
         String baseUrl,
