+++ src/main/java/com/lokoko/global/auth/service/AuthService.java
@@ -35,14 +35,14 @@ import com.lokoko.global.auth.jwt.exception.TokenInvalidException;
 import com.lokoko.global.auth.jwt.utils.CookieUtil;
 import com.lokoko.global.auth.jwt.utils.JwtExtractor;
 import com.lokoko.global.auth.jwt.utils.JwtProvider;
-import com.lokoko.global.auth.provider.google.GoogleOAuthClient;
-import com.lokoko.global.auth.provider.google.GoogleProperties;
+import com.lokoko.global.auth.provider.google.config.GoogleOAuthClient;
+import com.lokoko.global.auth.provider.google.config.GoogleProperties;
 import com.lokoko.global.auth.provider.google.dto.GoogleProfileDto;
 import com.lokoko.global.auth.provider.google.dto.GoogleTokenDto;
 import com.lokoko.global.auth.provider.google.dto.RoleUpdateResponse;
 import com.lokoko.global.auth.provider.google.dto.response.AfterLoginUserNameResponse;
-import com.lokoko.global.auth.provider.line.LineOAuthClient;
-import com.lokoko.global.auth.provider.line.LineProperties;
+import com.lokoko.global.auth.provider.line.config.LineOAuthClient;
+import com.lokoko.global.auth.provider.line.config.LineProperties;
 import com.lokoko.global.auth.provider.line.dto.LineProfileDto;
 import com.lokoko.global.auth.provider.line.dto.LineTokenDto;
 import com.lokoko.global.auth.provider.line.dto.LineUserInfoDto;
